requirements.txt
numpy==1.14.2
tensorflow-gpu==1.11.0
scipy==1.1.0